Update locales-zh-TW.xml
I'm done. 

If this goes, I shall be running the file through cconv to generate a simplified Chinese version. 

------------------------------------------

Another note to add about the punctuation setting. From the example given by the wiki, I can see that it is meant for title citation. The case for Chinese papers it that we use a special title mark 《》for books and a separate one〈〉for chapters and articles. As far as I know, we don't differentiate between outer quotes and inner quotes. This is the case for Taiwan. 

As for China, I think they use《》for outer quote and〈〉for inner quote without differentiating between books and chapters. In this case, the translated quotation marks「」and『』would do well to be swapped with these.