Update locales-zh-TW.xml
Resolved the ordinal issue. 

--------------------------------------------------------

Though more familiar with the Taiwanese convention, I believe differences in the Mainland would be minor, if any at all. Many Mainland scholars visit Taiwan to present papers, and vice-versa. 

I thought it is more practical to come up with a complete locale for both scripts first, and make changes later, if need be. At very least, the locales would be deploy-able. We'll never really know what changes are imperative until we start using the scripts. 

In any case, there is no single standard yet as to which terms to use or not to use in Chinese citations. It mostly depends on the author and the editor he is working with. The translation for online,「在線」, for example, can also be rendered as 「線上」, depending on context. 

The better journals would set a rough guideline for authors to follow and use a full-time editor to correct the format and diction on behalf of the author. The rest just leave it to the author to set and stick to his own standard. (And most authors don't really bother. )

The development of good and user-friendly Chinese csls would, hopefully, help set standards to which most people would gladly adhere to, if out of convenience. This would definitely save many long hours of editorial work, deciding on a new standard every time a new publication is in order, and manually editing citation formats to fit the newly-set standard.