Update locales-ar.xml
Changed delimiter in dates from Latin comma (,) to Arabic comma (،)
Changed tense of roles from modified to _marfu'_ (The form they were previously in - ending in ين- is used when that word has been modified by another word or is the object of a verb. The ون form is what it would take if it has been unmodified or at the beginning of a sentence - which is what the use case would be here.)
Added _kashida_ to short forms which have been abbreviated. 
Changed pages (single and multiple) to صـ (I don't think ص.ص. might be appropriate though it is analogous to pp.)
Completed a few entries. (They might not be totally in agreement with what is commonly used, but they can be modified in individual styles according to local tastes or changed here. But I think minimizing the mixing of LTR and RTL characters is very important...it can lead to serious confusion!)