Update locales-ar.xml
omitted the short forms, because I feel they are not needed, with regards to @gotoayman 's observation that rusoom (the plain noun form) is a better alternative to tarseem (a verbal noun form), and because they are in many cases redundant. If they are required by the processor i will re add them.