Translate original-work-published
Hi! I am in the process of contributing to the CSL community for the first time. A few days ago, I proposed a change to the APA style, which was replacing the text “Original work published” with the correspondent term, so it can be shown in many languages. My proposal to the APA style is under consideration. @adam3smith kindly suggested me to take a look here at Locales. I am entering to this particular locale because Spanish is my first language, and es-MX is the locale I use in my software. My translation of the text is:  “obra original publicada en”.  I hope you find this translation useful. If this is good for approval, then my plan is to submit the same translation to es-CL and es-ES next. Thanks for considering this proposal!