Update locales-ca-AD.xml
The "interviewed by" form is related to the author(s), not to the cited work. In Catalan this form has to agree in gender and number with the author(s), and "entrevistat per" would only agree with a single male author. The proposed form "interview made by" is gender neutral in Catalan and works with single or multiple authors.