Update locales-hr-HR.xml (#208)
* Update locales-hr-HR.xml

One of the Croatian translation of "et al." is "*i ostali*" but i haven't seen people using it. Instead, usually the term "*i suradnici*" (abbrev. "*i sur.*"; translated as "and collaborators") is used or rarely "*i drugi*" (abbrev. "*i dr.*"; translated as "and others") or even also rarely the term "et al." could be used as its origin is Latin. It would be great if all the terms could be used, so that in the csl file one can specify the desired term.

* Update locales-hr-HR.xml

---------

Co-authored-by: Brenton M. Wiernik <bwiernik@users.noreply.github.com>