en: Provide plural of `guest` in `verb` form (#360)
* en: Fix ungrammatical 'with guest' as a verb

Since 'with guest' is not a verb, rendering the verbal form simply as 'with' ensures grammatical results with an item including more than one `guest`. I cannot find the source for the phrase 'with guest' (it does not occur in any of APA, Chicago, MHRA, MLA, or New Hart's Rules); 'with' is a safer default.

* Use `<multiple>` for guests